backend/requirements.txt
fastapi[all]>=0.115.12
langchain>=0.3.25
langchain-openai>=0.3.18
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
sqlalchemy>=2.0.41
uvicorn>=0.34.2
fastapi
uvicorn[standard]
